Reorganize: Shift location of project structure
Originally each project was going to be managed separately. But as I built the 1st, it became a new showcase website of my abilities. So each project is a page within the website. Meaning that I should combine the project.

Then keep each Projects JS files separate while it is under review.